+++ src/main/java/com/cathalob/medtracker/repository/PractitionerRoleRequestRepository.java
@@ -3,5 +3,5 @@ package com.cathalob.medtracker.repository;
 import com.cathalob.medtracker.model.PractitionerRoleRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PractitionerRoleRequestRepository extends JpaRepository<PractitionerRoleRequest, Integer> {
+public interface PractitionerRoleRequestRepository extends JpaRepository<PractitionerRoleRequest, Long> {
 }
